pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fiap-healthcare-plus-mlops"
version = "0.5.0"
description = "Pipeline reproducivel de MLOps para triagem cardiaca da FIAP HealthCare Plus."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "flask>=3.1.0,<4.0.0",
  "joblib>=1.4.0,<2.0.0",
  "mlflow>=2.20.0,<3.0.0",
  "numpy>=2.2.0,<3.0.0",
  "pandas>=2.2.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "scikit-learn>=1.6.0,<2.0.0",
]

[dependency-groups]
dev = [
  "mypy>=1.15.0,<2.0.0",
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.11.0,<0.12.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.13"
packages = ["src"]
ignore_missing_imports = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
